+++ mulang.cabal
@@ -49,6 +49,7 @@ library
     Language.Mulang.Inspector.Logic
     Language.Mulang.Inspector.Procedural
     Language.Mulang.Inspector.Combiner
+    Language.Mulang.Inspector.Generalized
     Language.Mulang.Inspector.Generic.Duplication
     Language.Mulang.Inspector.Generic.Smell
     Language.Mulang.Parsers
