+++ src/Language/Mulang/Inspector.hs
@@ -3,6 +3,7 @@ module Language.Mulang.Inspector (
   IdentifierInspection,
   module Language.Mulang.Inspector.Generic,
   module Language.Mulang.Inspector.Combiner,
+  module Language.Mulang.Inspector.Generalized,
   module Language.Mulang.Inspector.Typed,
   module Language.Mulang.Inspector.ObjectOriented,
   module Language.Mulang.Inspector.ObjectOriented.Polymorphism,
@@ -14,6 +15,7 @@ import Language.Mulang.Inspector.Primitive (Inspection, IdentifierInspection)
 import Language.Mulang.Inspector.Generic
 import Language.Mulang.Inspector.Typed
 import Language.Mulang.Inspector.Combiner
+import Language.Mulang.Inspector.Generalized
 import Language.Mulang.Inspector.ObjectOriented
 import Language.Mulang.Inspector.ObjectOriented.Polymorphism
 import Language.Mulang.Inspector.Functional
